Exclude some signals/syscall code on Windows
