+++ events.go
@@ -8,11 +8,11 @@ func main() {
   go WatchSignals(finished)
 
   RegisterEventHandler("sigint",func(){
-    fmt.Prinln("got signint") 
+    fmt.Prinln("got signint")
   })
 
   // block here, should put this goroutine into sleep mode
-  v := <-finished 
+  v := <-finished
 
 }
 */
@@ -21,9 +21,7 @@ package gou
 
 import (
 	"os"
-	"os/signal"
 	"sync"
-	"syscall"
 	"time"
 )
 
@@ -33,7 +31,7 @@ var (
 )
 
 // Wait for condition (defined by func) to be true
-// this is mostly for testing, but a utility to 
+// this is mostly for testing, but a utility to
 // create a ticker checking back every 100 ms to see
 // if something (the supplied check func) is done
 //
@@ -54,64 +52,6 @@ func WaitFor(check func() bool, timeoutSecs int) {
 			break
 		}
 		tryct++
-	}
-}
-
-// a watcher that tries to trap sys signals so you can gracefully shutdown.  
-// Make sure to only call this once.  Currently monitors events [sigterm,
-// sigint, sigabrt,sigquit,sigstop, sigusr1,sigusr2]
-func WatchSignals(quit chan bool) {
-	var sig os.Signal
-	sigIn := make(chan os.Signal, 1)
-	signal.Notify(sigIn)
-
-	defer func() {
-		if r := recover(); r != nil {
-			Debug("Recovered in Watch Signals", r)
-		}
-	}()
-
-	for {
-		sig = <-sigIn
-		if sig.(os.Signal) == syscall.SIGTERM {
-			Log(DEBUG, "in WatchSignal Handle Sig term")
-			RunEventHandlers("sigterm")
-			RunEventHandlers("onexit")
-			quit <- true
-		}
-		if sig.(os.Signal) == syscall.SIGINT {
-			Log(DEBUG, "in WatchSignal Handle Sig Int")
-			RunEventHandlers("sigint")
-			RunEventHandlers("onexit")
-			quit <- true
-		}
-		if sig.(os.Signal) == syscall.SIGABRT {
-			Log(DEBUG, "in WatchSignal Handle SIGABRT")
-			RunEventHandlers("sigabrt")
-			RunEventHandlers("onexit")
-			quit <- true
-			//os.Exit(9)
-		}
-		if sig.(os.Signal) == syscall.SIGQUIT {
-			Log(DEBUG, "in WatchSignal Handle SIGQUIT")
-			RunEventHandlers("sigquit")
-			RunEventHandlers("onexit")
-			quit <- true
-		}
-		if sig.(os.Signal) == syscall.SIGSTOP {
-			Log(DEBUG, "in WatchSignal Handle SIGSTOP")
-			RunEventHandlers("SIGSTOP")
-		}
-		if sig.(os.Signal) == syscall.SIGUSR1 {
-			Log(DEBUG, "in WatchSignal Handle SIGUSR1")
-			RunEventHandlers("SIGUSR1")
-		}
-		if sig.(os.Signal) == syscall.SIGUSR2 {
-			Log(DEBUG, "in WatchSignal Handle SIGUSR2")
-			RunEventHandlers("SIGUSR2")
-		}
-
-		Debug(sig)
 	}
 }
