+++ log.go
@@ -5,8 +5,6 @@ import (
 	"log"
 	"os"
 	"strings"
-	"syscall"
-	"unsafe"
 )
 
 const (
@@ -243,24 +241,3 @@ type winsize struct {
 const (
 	_TIOCGWINSZ = 0x5413 // OSX 1074295912
 )
-
-// Determine is this process is running in a Terminal or not?
-func IsTerminal() bool {
-	ws := &winsize{}
-	isTerm := true
-	defer func() {
-		if r := recover(); r != nil {
-			isTerm = false
-		}
-	}()
-	// This blows up on windows
-	retCode, _, _ := syscall.Syscall(syscall.SYS_IOCTL,
-		uintptr(syscall.Stdin),
-		uintptr(_TIOCGWINSZ),
-		uintptr(unsafe.Pointer(ws)))
-
-	if int(retCode) == -1 {
-		return false
-	}
-	return isTerm
-}
